+++ packages/core/src/components/MapContainer.vue
@@ -161,12 +161,15 @@ export default Vue.extend({
       this.checkServiceAvailability()
     }
     addEventListener('resize', this.updateMaxMobileHeight)
+    addEventListener('resize', this.updateHasSmallDisplay)
+    this.updateHasSmallDisplay()
   },
   beforeDestroy() {
     removeEventListener('resize', this.updateMaxMobileHeight)
+    removeEventListener('resize', this.updateHasSmallDisplay)
   },
   methods: {
-    ...mapMutations(['setMap', 'setConfiguration']),
+    ...mapMutations(['setConfiguration', 'setHasSmallDisplay', 'setMap']),
     ...mapActions([
       'updateDragAndZoomInteractions',
       'useExtendedMasterportalapiMarkers',
@@ -208,6 +211,12 @@ export default Vue.extend({
             })
             .catch(console.error)
         )
+    },
+    updateHasSmallDisplay() {
+      this.setHasSmallDisplay(
+        window.innerHeight <= SMALL_DISPLAY_HEIGHT ||
+          window.innerWidth <= SMALL_DISPLAY_WIDTH
+      )
     },
     updateListeners(hasWindowSize: boolean) {
       const mapContainer = this.$refs['polar-map-container']
@@ -302,6 +311,11 @@ export default Vue.extend({
 .polar-shadow {
   height: 100%;
   width: 100%;
+}
+
+.v-tooltip__content {
+  background-color: #595959;
+  border: 2px solid #fff;
 }
 
 .v-application {