+++ packages/plugins/GeoLocation/src/components/GeoLocation.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <v-tooltip class="polar-geo-location-wrap" left>
+    <v-tooltip left :disabled="hasSmallDisplay">
       <template #activator="{ on, attrs }">
         <v-btn
           :class="{
@@ -29,6 +29,7 @@ import { mapActions, mapGetters } from 'vuex'
 export default Vue.extend({
   name: 'GeoLocation',
   computed: {
+    ...mapGetters(['hasSmallDisplay']),
     ...mapGetters('plugin/geoLocation', [
       'geolocation',
       'renderType',
@@ -51,9 +52,4 @@ export default Vue.extend({
 })
 </script>
 
-<style lang="scss" scoped>
-.polar-geo-location-wrap {
-  display: flex;
-  flex-direction: column;
-}
-</style>
+<style lang="scss" scoped></style>