+++ packages/plugins/Gfi/src/store/actions.ts
@@ -2,9 +2,12 @@ import debounce from 'lodash.debounce'
 import { Coordinate } from 'ol/coordinate'
 import { Feature as GeoJsonFeature } from 'geojson'
 import { Style, Fill, Stroke } from 'ol/style'
+import Overlay from 'ol/Overlay'
 import { GeoJSON } from 'ol/format'
 import { rawLayerList } from '@masterportal/masterportalapi/src'
 import { PolarActionTree } from '@polar/lib-custom-types'
+import { getTooltip, Tooltip } from '@polar/lib-tooltip'
+import { Feature } from 'ol'
 import {
   featureDisplayLayer,
   clear,
@@ -15,6 +18,8 @@ import { GfiGetters, GfiState } from '../types'
 import sortFeatures from '../utils/sortFeatures'
 
 const writer = new GeoJSON()
+
+let debouncedVisibilityChangeIndicator
 
 const actions: PolarActionTree<GfiState, GfiGetters> = {
   /**
@@ -54,6 +59,82 @@ const actions: PolarActionTree<GfiState, GfiGetters> = {
         ),
       })
     )
+
+    dispatch('setupTooltip')
+    dispatch('setupFeatureVisibilityUpdates')
+  },
+  setupTooltip({ getters: { gfiConfiguration }, rootGetters: { map } }) {
+    const tooltipLayerIds = Object.keys(gfiConfiguration.layers).filter(
+      (key) => gfiConfiguration.layers[key].showTooltip
+    )
+    if (!tooltipLayerIds.length) {
+      return
+    }
+
+    let element: Tooltip['element'], unregister: Tooltip['unregister']
+    const overlay = new Overlay({
+      positioning: 'bottom-center',
+      offset: [0, -5],
+    })
+    map.addOverlay(overlay)
+    map.on('pointermove', ({ pixel, dragging }) => {
+      if (dragging) {
+        return
+      }
+      let hasFeatureAtPixel = false
+      // stops on return `true`, thus only using the uppermost feature
+      map.forEachFeatureAtPixel(
+        pixel,
+        (feature, layer) => {
+          if (!(feature instanceof Feature)) {
+            return false
+          }
+          hasFeatureAtPixel = true
+
+          overlay.setPosition(map.getCoordinateFromPixel(pixel))
+
+          if (unregister) {
+            unregister()
+          }
+          ;({ element, unregister } = getTooltip({
+            localeKeys:
+              // @ts-expect-error | it exists by virtue of layerFilter below
+              gfiConfiguration.layers[layer.get('id')].showTooltip(feature),
+          }))
+          overlay.setElement(element)
+          return true
+        },
+        {
+          layerFilter: (layer) => tooltipLayerIds.includes(layer.get('id')),
+        }
+      )
+      if (!hasFeatureAtPixel) {
+        overlay.setPosition(undefined)
+      }
+    })
+  },
+  setupFeatureVisibilityUpdates({ commit, state, getters, rootGetters }) {
+    // debounce to prevent update spam
+    debouncedVisibilityChangeIndicator = debounce(
+      () =>
+        commit(
+          'setVisibilityChangeIndicator',
+          state.visibilityChangeIndicator + 1
+        ),
+      10
+    )
+    const usedLayers = Object.keys(getters.gfiConfiguration.layers)
+    rootGetters.map
+      .getLayers()
+      .getArray()
+      .forEach((layer) => {
+        if (usedLayers.includes(layer.get('id'))) {
+          layer
+            // @ts-expect-error | layers reaching this have a source
+            .getSource()
+            .on('changefeature', debouncedVisibilityChangeIndicator)
+        }
+      })
   },
   close({ commit, dispatch }) {
     commit('clearFeatureInformation')