+++ packages/plugins/Gfi/src/types.ts
@@ -35,6 +35,8 @@ export interface GfiState {
   imageLoaded: boolean
   /** currently visible feature index regarding  */
   visibleWindowFeatureIndex: number
+  /** count up if something in relevant features changed */
+  visibilityChangeIndicator: number
   /** default style for stroke and fill of the highlighted feature. */
   defaultHighlightStyle: HighlightStyle
   page: number