+++ packages/types/custom/core.ts
@@ -189,6 +189,38 @@ export interface ExportConfiguration extends PluginOptions {
   showPng?: boolean
 }
 
+export interface FilterConfigurationTimeOption {
+  amounts: number[]
+  unit: 'days'
+}
+
+interface FilterConfigurationTime {
+  targetProperty: string
+  pattern?: string
+  last?: FilterConfigurationTimeOption[]
+  next?: FilterConfigurationTimeOption[]
+  freeSelection?: {
+    now?: 'until' | 'from'
+    unit: 'days'
+  }
+}
+
+interface FilerConfigurationCategory {
+  selectAll?: boolean
+  targetProperty: string
+  knownValues: (string | number)[]
+}
+
+export interface FilterConfiguration extends PluginOptions {
+  layers: Record<
+    string,
+    {
+      categories?: FilerConfigurationCategory[]
+      time?: FilterConfigurationTime
+    }
+  >
+}
+
 /** Configuration of GFI feature regarding a specific layer */
 export interface GfiLayerConfiguration {
   /**
@@ -227,6 +259,7 @@ export interface GfiLayerConfiguration {
   geometryName?: string
   // format the response is known to come in (e.h. "GML")
   format?: string
+  showTooltip?: (feature: Feature) => [string, string][]
 }
 
 export type BoundaryOnError = 'strict' | 'permissive'
@@ -520,6 +553,7 @@ export interface MapConfig {
   attributions?: AttributionsConfiguration
   draw?: DrawConfiguration
   export?: ExportConfiguration
+  filter?: FilterConfiguration
   fullscreen?: FullscreenConfiguration
   geoLocation?: GeoLocationConfiguration
   gfi?: GfiConfiguration
@@ -569,7 +603,9 @@ export interface CoreState {
   components: number
   configuration: MapConfig
   errors: PolarError[]
+  hasSmallDisplay: boolean
   hovered: number
+  language: string
   map: number
   moveHandle: number
   plugin: object