+++ packages/plugins/Attributions/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## unpublished
 
-- Feature: Add link to repository and license as required by EUPL v1.2
+- Feature: Add link to repository and license as required by EUPL v1.2.
+- Feature: Include title in AttributionContent so users receive this information on all devices.
+- Fix: Attributions are now being translated when the window is open
 
 ## 1.0.0