+++ packages/plugins/Fullscreen/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Hide obstructive tooltip on small devices.
+- Fix: Documentation error regarding plugin state.
+
 ## 1.0.0
 
 Initial release.