+++ packages/plugins/LayerChooser/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Hide obstructive tooltip on small devices.
+
 ## 1.0.0
 
 Initial release.