+++ packages/types/custom/CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## unpublished
 
+- Feature: Add currently active language key as parameter `language` to `CoreState `.
+- Feature: Add types for new plugin `Filter`, namely `FilterConfiguration`; also to `MapConfig` as configurable plugin.
 - Feature: Add types for new core state and getters. (`zoomLevel`, `hovered`, `selected`)
 - Feature: Add parameters `renderType`, `featureList`, and `activeLayerPath` to `GfiConfiguration`.
 - Feature: Add type `FeatureList` for `GfiConfiguration`.
+- Feature: Add parameter `showTooltip` to `GfiLayerConfiguration`.
 - Feature: Add parameter `renderType` to `GeoLocationConfiguration`.
 - Feature: Add export of `RenderType`.
 - Feature: Add parameter `focusAfterSearch` to `AddressSearchConfiguration`.
+- Feature: Add parameter `hasSmallDisplay` to `CoreState`.
 
 ## 1.1.0